more modularization test for health_controller.py and guild_controller.py
