+++ bot/__init__.py
@@ -1,0 +1,2 @@
+from .TS3Auth import AuthRequest
+from .TS3Bot import Bot
