+++ bot/__main__.py
@@ -2,10 +2,10 @@ import logging
 import sys
 
 from bot.main import main
-from bot.util.logging import initialize
+from bot.util import initialize_logging
 
 LOG = logging.getLogger(__name__)
 
 if __name__ == '__main__':
-    initialize()
+    initialize_logging()
     main(sys.argv[1:])
