+++ bot/messages/__init__.py
@@ -1,0 +1,2 @@
+from .bot_messages import get_locale
+from .locale import Locale
