+++ bot/rest/controller/__init__.py
@@ -1,0 +1,3 @@
+from .guild_controller import GuildController
+from .health_controller import HealthController
+from .other_controller import OtherController
