+++ bot/ts/__init__.py
@@ -1,0 +1,1 @@
+from .TS3Facade import Channel, TS3Facade
