+++ bot/util/__init__.py
@@ -1,0 +1,2 @@
+from .StringShortener import StringShortener
+from .logging import initialize_logging
