+++ bot/util/logging.py
@@ -4,15 +4,16 @@ import sys
 FORMAT = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
 
 
-def initialize(file="./ts3bot.log", level: int = logging.DEBUG):
+def initialize_logging(file="./ts3bot.log", level: int = logging.DEBUG):
     formatter = logging.Formatter(FORMAT)
 
     handlers = []
-    for handler, level in [(logging.StreamHandler(sys.stdout), logging.DEBUG), (logging.FileHandler(file, delay=True), logging.DEBUG)]:
-        handler.setLevel(level)
+    for handler, handler_level in [(logging.StreamHandler(sys.stdout), logging.DEBUG), (logging.FileHandler(file, delay=True), logging.DEBUG)]:
+        handler.setLevel(handler_level)
         handler.setFormatter(formatter)
         handlers.append(handler)
 
+    # noinspection PyArgumentList
     logging.basicConfig(
         level=level,
         format=FORMAT,
